# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Esiana Core API
  version: 0.9.0
  description: |
    REST API for Esiana campaign narrative infrastructure.

    **Authentication:** Session cookie (`esiana_token`) or Bearer API token with scopes.

    **Campaign scope:** Most lore routes live under `/api/campaigns/{campaignHandle}/`.

    ## Not available in core (v1.0)

    - **Search** — no unified `/search` REST endpoint; use wiki hubs, link-index, or plugin search.
    - **Webhooks** — outbound delivery queue deferred post-1.0.

    ## Human documentation

    API guides and architecture explainers live in the [Esiana docs wiki](https://github.com/Esiana-ttrpg/docs):
    [API overview](../../docs/api/overview.md), [campaign model](../../docs/architecture/campaign-model.md),
    [plugin development](../../docs/plugin-development/getting-started.md).

    **Interactive explorer:** open `/api/docs` on your running Esiana instance (version-locked to this release).
servers:
  - url: http://localhost:3001
    description: Local backend
...
paths:
...
  /api/campaigns:
    get:
      tags: [Campaigns]
      summary: List campaigns for current user
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Campaign list
          content:
            application/json:
              schema:
                type: object
                properties:
                  campaigns:
                    type: array
                    items:
                      $ref: '#/components/schemas/CampaignSummary'
              example:
                campaigns:
                  - id: cmp_01ABC
                    name: Port Haven
                    handle: port-haven
                    role: GAMEMASTER
                    isMember: true
                    isCampaignOwner: true
                    heroImageUrl: null
                    isArchived: false
    post:
      tags: [Campaigns]
      summary: Create campaign (wizard)
      description: Multipart form — cover image, Obsidian ZIP, sovereign backup ZIP, or calendar JSON optional.
      security:
        - cookieAuth: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                handle:
                  type: string
                markdownZipFile:
                  type: string
                  format: binary
                backupZipFile:
                  type: string
                  format: binary
      responses:
        '201':
          description: Campaign created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignCreated'
              example:
                campaign:
                  id: cmp_01ABC
                  name: Port Haven
                  handle: port-haven
                  role: GAMEMASTER
                  isMember: true
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Campaign name is required
  /api/campaigns/{campaignHandle}:
    patch:
      tags: [Campaigns]
      summary: Update campaign settings
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/campaignHandle'
      responses:
        '200':
          description: Updated campaign
...
  /api/import-providers:
    get:
      tags: [Import]
      summary: List import providers (core + plugins)
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Provider descriptors
          content:
            application/json:
              schema:
                type: object
                required: [core, plugins]
                properties:
                  core:
                    type: array
                    items:
                      $ref: '#/components/schemas/CoreImportProvider'
                  plugins:
                    type: array
                    items:
                      $ref: '#/components/schemas/PluginImportProvider'
              example:
                core:
                  - source: core
                    id: markdown-zip
                    label: Markdown ZIP
                    description: Obsidian-style markdown archive
                  - source: core
                    id: esiana-backup
                    label: Esiana Backup
                    description: Full campaign backup package
                plugins:
                  - source: plugin
                    id: obsidian-vault
                    pluginId: obsidian
                    providerId: vault-import
                    label: Vault Import
                    description: ''
                    requiresFile: false
                    requiresAuth: false
  /api/plugins:
    get:
      tags: [Plugins]
      summary: List installed system plugins
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Plugin catalog
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: esiana_token
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API token with scopes
  parameters:
    campaignHandle:
      name: campaignHandle
      in: path
      required: true
      schema:
        type: string
      example: port-haven
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    CampaignSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        handle:
          type: string
        role:
          type: string
          nullable: true
        isMember:
          type: boolean
        isCampaignOwner:
          type: boolean
        heroImageUrl:
          type: string
          nullable: true
        isArchived:
          type: boolean
    CampaignCreated:
      type: object
      properties:
        campaign:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            handle:
              type: string
            role:
              type: string
            isMember:
              type: boolean
...
    CoreImportProvider:
      type: object
      required: [source, id, label, description]
      properties:
        source:
          type: string
          enum: [core]
        id:
          type: string
        label:
          type: string
        description:
          type: string
    PluginImportProvider:
      type: object
      required: [source, id, pluginId, providerId, label, description, requiresFile, requiresAuth]
      properties:
        source:
          type: string
          enum: [plugin]
        id:
          type: string
        pluginId:
          type: string
        providerId:
          type: string
        label:
          type: string
        description:
          type: string
        requiresFile:
          type: boolean
        requiresAuth:
          type: boolean